Prediction values are always an integer between 0 and 100.

+++ footystats-tools-backend/src/main/java/de/ludwig/footystats/tools/backend/services/prediction/BetPredictionDistribution.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class BetPredictionDistribution {
 	@Getter
 	@Setter
-	private Float predictionPercent;
+	private Integer predictionPercent;
 	@Getter
 	@Setter
 	private Long count;
