+++ footystats-tools-backend/src/main/java/de/ludwig/footystats/tools/backend/services/prediction/PredictionResult.java
@@ -2,7 +2,7 @@ package de.ludwig.footystats.tools.backend.services.prediction;
 
 import java.util.List;
 
-public record PredictionResult (Float betSuccessInPercent,
+public record PredictionResult (Integer betSuccessInPercent,
                                 boolean betOnThis,
                                 PredictionAnalyze analyzeResult,
                                 List<InfluencerResult> influencerDetailedResult){
